windows build cleanup
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@9206 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ src/mod/formats/mod_shout/decode_ntom.c
@@ -18,7 +18,7 @@
 #define WRITE_SAMPLE(samples,sum,clip) \
   if( (sum) > 32767.0) { *(samples) = 0x7fff; (clip)++; } \
   else if( (sum) < -32768.0) { *(samples) = -0x8000; (clip)++; } \
-  else { *(samples) = sum; }
+  else { *(samples) = (short)sum; }
 
 
 //static unsigned long ntom_val[2] = { NTOM_MUL >> 1, NTOM_MUL >> 1 };
